moved all database logic part to separate module
