+++ core/src/main/java/com/krypton/cloud/CloudApplication.java
@@ -3,11 +3,10 @@ package com.krypton.cloud;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"com.krypton.cloud", "com.krypton.databaselayer"})
 public class CloudApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(CloudApplication.class, args);
 	}
-
 }
