--- core/src/main/java/com/krypton/cloud/repository/FileRepository.java
@@ -1,16 +1,0 @@
-package com.krypton.cloud.repository;
-
-import com.krypton.cloud.model.File;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface FileRepository extends JpaRepository<File, UUID> {
-
-    @Query(value = "select * from file where file.path = :path", nativeQuery = true)
-    File getByPath(@Param("path") String path);
-}
