--- core/src/main/java/com/krypton/cloud/service/record/IOEntityRecordService.java
@@ -1,7 +1,0 @@
-package com.krypton.cloud.service.record;
-
-public interface IOEntityRecordService<T> extends RecordService<T> {
-
-    T getByPath(String path);
-
-}
