+++ core/src/main/kotlin/com/krypton/cloud/config/WebConfig.kt
@@ -7,10 +7,10 @@ import org.springframework.web.reactive.config.WebFluxConfigurationSupport
 @Configuration
 open class WebConfig : WebFluxConfigurationSupport() {
 
-    public override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-                .allowedMethods("HEAD", "GET", "PUT", "POST", "DELETE", "PATCH")
-                .allowedOrigins("*")
-                .allowedHeaders("*")
-    }
+	override fun addCorsMappings(registry: CorsRegistry) {
+		registry.addMapping("/**")
+			.allowedMethods("HEAD", "GET", "PUT", "POST", "DELETE", "PATCH")
+			.allowedOrigins("*")
+			.allowedHeaders("*")
+	}
 }
