--- core/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/cloudFilesRecord?useSSL=false&serverTimezone=UTC&useLegacyDatetimeCode=false&allowPublicKeyRetrieval=true
-spring.datasource.username=krypt0n
-spring.datasource.password=1708
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
-spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=create
-spring.output.ansi.enabled=always
-multipart.enabled=false
-server.port=8080
