+++ core/src/main/resources/application.yml
@@ -1,0 +1,24 @@
+spring:
+  datasource:
+    url: "jdbc:mysql://127.0.0.1:3306/cloudFilesRecord?useSSL=false&serverTimezone=UTC&useLegacyDatetimeCode=false&allowPublicKeyRetrieval=true"
+    username: "krypt0n"
+    password: 1708
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  jpa:
+    generate-ddl: true
+    properties:
+      hibernate:
+        dialect: "org.hibernate.dialect.MySQL5InnoDBDialect"
+    show-sql: true
+    hibernate:
+      ddl-auto: create-drop
+  output:
+    ansi:
+      enabled: always
+  servlet:
+    multipart:
+      enabled: false
+  main:
+    allow-bean-definition-overriding: true
+server:
+  port: 8080
