+++ database-layer/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'database-layer'
+
+include ":common", ":util"
