+++ database-layer/src/main/java/com/krypton/databaselayer/service/IOEntityRecordService.java
@@ -1,0 +1,7 @@
+package com.krypton.databaselayer.service;
+
+public interface IOEntityRecordService<T> extends RecordService<T> {
+
+    T getByPath(String path);
+
+}
