+++ database-layer/src/test/java/com/krypton/databaselayer/DatabaseLayerApplicationTests.java
@@ -1,0 +1,16 @@
+package com.krypton.databaselayer;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class DatabaseLayerApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
